splitted up PackedPixelTests into one test file for each pixel format
